revert to testfactory.go
