+++ internal/cli/test/test.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"github.com/kyma-incubator/reconciler/internal/cli"
 	"github.com/kyma-incubator/reconciler/internal/persistency"
+	"github.com/kyma-incubator/reconciler/pkg/db"
 	"github.com/kyma-incubator/reconciler/pkg/test"
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/require"
@@ -18,7 +19,7 @@ func NewTestOptions(t *testing.T) *cli.Options {
 	}
 
 	var err error
-	cliOptions.Registry, err = persistency.NewRegistry(test.NewTestConnectionFactory(t), true)
+	cliOptions.Registry, err = persistency.NewRegistry(db.NewTestConnectionFactory(t), true)
 	require.NoError(t, err)
 
 	return cliOptions
