+++ pkg/db/testfactory.go
@@ -1,0 +1,52 @@
+package db
+
+import (
+	"log"
+	"testing"
+
+	"github.com/kyma-incubator/reconciler/pkg/test"
+	"github.com/stretchr/testify/require"
+)
+
+func NewTestConnectionFactory(t *testing.T) ConnectionFactory {
+	configFile, err := test.GetConfigFile()
+	require.NoError(t, err)
+
+	connFac, err := NewConnectionFactory(configFile, false, true)
+	require.NoError(t, err)
+
+	require.NoError(t, connFac.Init(false))
+	return connFac
+}
+
+func NewTestConnection(t *testing.T) Connection {
+	connFac := NewTestConnectionFactory(t)
+	conn, err := connFac.NewConnection()
+	require.NoError(t, err)
+	return conn
+}
+
+func CleanUpTables(t *testing.T) {
+	dbConn := NewTestConnection(t)
+	log.Println("after tests")
+	queryPrefix := "TRUNCATE TABLE "
+	if dbConn.Type() == SQLite {
+		queryPrefix = "DELETE FROM "
+	}
+	tableNames := [...]string{
+		"config_values",
+		"config_keys",
+		"config_cachedeps",
+		"config_cache",
+		"scheduler_operations",
+		"scheduler_reconciliations",
+		"inventory_clusters",
+		"inventory_cluster_configs",
+		"inventory_cluster_config_statuses",
+	}
+	for _, tableName := range tableNames {
+		_, err := dbConn.DB().Exec(queryPrefix + tableName)
+		require.NoError(t, err)
+	}
+
+}
