+++ pkg/kv/repository_test.go
@@ -2,7 +2,6 @@ package kv
 
 import (
 	"fmt"
-	"github.com/kyma-incubator/reconciler/pkg/test"
 	"testing"
 	"time"
 
@@ -338,7 +337,7 @@ func TestRepositoryValues(t *testing.T) {
 }
 
 func newKeyValueRepo(t *testing.T) *Repository {
-	ceRepo, err := NewRepository(test.NewTestConnection(t), true)
+	ceRepo, err := NewRepository(db.NewTestConnection(t), true)
 	require.NoError(t, err)
 	return ceRepo
 }
